"added authentication"

+++ main.py
@@ -1,15 +1,162 @@
 from flask import Flask, render_template, request, redirect, jsonify, url_for, \
     flash
-from os import environ
+from flask import session as login_session
+import random, string
 app = Flask(__name__)
 from sqlalchemy import create_engine, asc, desc
 from sqlalchemy.orm import sessionmaker
-from db_setup import Base, Category, Item
+from db_setup import Base, Category, Item, User
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+import logging
 
 engine = create_engine('sqlite:///itemcatalog.db')
 Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+
+#login route
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)for
+                    x in xrange(32))
+    login_session['state'] = state
+    #return "The current session state is %s " % login_session['state']
+    return render_template('login.html', STATE=state)
+
+#gconnect
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_credentials = login_session.get('credentials')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_credentials is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['credentials'] = credentials
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+    login_session['email'] = data['email']
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['access_token'] = credentials.access_token
+    login_session['logged_in'] = True
+    user_id = getUserID(data['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+#logout user
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session['access_token']
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: '
+    print login_session['username']
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % \
+          login_session['access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        login_session['logged_in'] = False
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        flash('You are now logged out!')
+        return redirect(url_for('showCategories'))
+        #return response
+    else:
+
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 
 # get name of category
 @app.context_processor
@@ -60,10 +207,34 @@ def categoryJSON(category_id):
     item = session.query(Item).filter_by(id=category_id).one()
     return jsonify(Category=[item.serialize])
 
+# User Helper Functions
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
 # Create a new category
 @app.route('/catalog/category/add', methods=['GET', 'POST'])
 def addCategory():
     #To do: check if a registered user is accessing the page, set cookies
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         try:
             category = session.query(Category).filter_by(name=request.form[
@@ -73,20 +244,26 @@ def addCategory():
                 return redirect(url_for('showCategories'))
         except:
             newCategory = Category(name=request.form['name'],
-                               description = request.form['description'])
+                               description = request.form['description'],
+                                   user_id=login_session['user_id'])
             session.add(newCategory)
             flash('New category %s Successfully created')
             session.commit()
             return redirect(url_for('showCategories'))
     else:
-        return render_template('addCategory.html')
+        return render_template('addCategory.html', login_session=login_session)
 
 
 # Edit a categoryshowcategoriesshowCategories
 @app.route('/catalog/category/<int:category_id>/edit', methods=['GET', 'POST'])
 def editCategory(category_id):
-    #To do: check if user is logged in or not, set cookies
+    if 'username' not in login_session:
+        return redirect('/login')
     editedCategory = session.query(Category).filter_by(id=category_id).one()
+    if 'username' in login_session and login_session['user_id'] != \
+            editedCategory.user_id :
+        flash('Category is not owned by you and hence can not be edited')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         if request.form['name']:
             try:
@@ -104,21 +281,29 @@ def editCategory(category_id):
         flash('Category successfully edited %s' % editedCategory.name)
         return redirect(url_for('showCategories'))
     else:
-        return render_template('editCategory.html', category=editedCategory)
+        return render_template('editCategory.html', category=editedCategory,
+                               login_session=login_session)
 
 
 # Delete a category
 @app.route('/catalog/category/<int:category_id>/delete',
            methods=['GET', 'POST'])
 def deleteCategory(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     category = session.query(Category).filter_by(id=category_id).one()
+    if 'username' in login_session and login_session['user_id'] != \
+            category.user_id :
+        flash('Category is not owned by you and hence can not be deleted')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         session.delete(category)
         flash('%s Successfully deleted' % category.name)
         session.commit()
         return redirect(url_for('showCategories'))
     else:
-        return render_template('deleteCategory.html', category=deleteCategory)
+        return render_template('deleteCategory.html', category=deleteCategory,
+                               login_session=login_session)
 
 
 # Show all categories
@@ -127,39 +312,52 @@ def deleteCategory(category_id):
 def showCategories():
     categories = session.query(Category).order_by(asc(Category.id))
     items = session.query(Item).order_by(desc(Item.id)).limit(10)
-    return render_template('category.html', categories= categories, items=items)
+    return render_template('category.html', categories= categories,
+                           items=items, login_session=login_session)
 
 
 # Show details of a selected category
 @app.route('/catalog/category/<int:category_id>')
 def showSelectedCategory(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
-    return render_template('selectedCategory.html', category=category)
+    return render_template('selectedCategory.html', category=category,
+                           login_session=login_session)
 
 
 # Create an item
 @app.route('/catalog/category/item/add',
            methods=['GET', 'POST'])
 def addItem():
+    if 'username' not in login_session:
+        return redirect('/login')
+
     categories = session.query(Category).order_by(asc(Category.id))
     if request.method == 'POST':
         newItem = Item(name=request.form['name'],
                        description=request.form['description'],
-                       category_id=request.form['category'])
+                       category_id=request.form['category'],
+                       user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash('New item %s is successfully created' % newItem.name)
         return redirect(url_for('showCategories'))
         #return redirect(url_for('showItemsCategory', category_id=category_id))
     else:
-        return render_template('addItem.html', category_options=categories)
+        return render_template('addItem.html', category_options=categories,
+                               login_session=login_session)
 
 
 # Edit an item
 @app.route('/catalog/category/<int:category_id>/item/<int:item_id>/edit',
            methods=['GET', 'POST'])
 def editItem(category_id, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if 'username' in login_session and login_session['user_id'] != \
+            item.user_id :
+        flash('Item is not owned by you and hence can not be edited')
+        return redirect(url_for('showCategories'))
     categories = session.query(Category).order_by(asc(Category.id))
     if request.method == 'POST':
         if request.form['name']:
@@ -174,21 +372,29 @@ def editItem(category_id, item_id):
         return redirect(url_for('showCategories'))
     else:
         return render_template('editItem.html', category_id=category_id,
-                               item=item, category_options=categories)
+                               item=item, category_options=categories,
+                               login_session=login_session)
 
 
 # Delete an item
 @app.route('/catalog/category/<int:category_id>/item/<int:item_id>/delete',
            methods=['GET', 'POST'])
 def deleteItem(category_id, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if 'username' in login_session and login_session['user_id'] != \
+            item.user_id :
+        flash('Item is not owned by you and hence can not be deleted')
+        return redirect(url_for('showCategories'))
     if request.method == 'POST':
         session.delete(item)
         session.commit()
         flash('Item %s successfully deleted' % item.name)
         return redirect(url_for('showCategories'))
     else:
-        return render_template('deleteItem.html', item=item)
+        return render_template('deleteItem.html', item=item,
+                               login_session=login_session)
 
 
 # Show all items in a category
@@ -221,7 +427,4 @@ def recentItems():
 if __name__ == '__main__':
     app.secret_key = 'test_secret_key'
     app.debug = True
-    #port = int(os.environ.get('PORT', 5000))
-    #app.run(host='0.0.0.0', port=port)
-    #app.run(debug=False, port=environ.get("PORT", 5000), processes=2)
     app.run(host='0.0.0.0', port=5000)
